+++ ui/src/main/java/com/example/ui/ui/di/ViewModelsModule.kt
@@ -1,8 +1,6 @@
 package com.example.ui.ui.di
 
-
-
-import com.example.ui.ui.home.HomeViewModel
+import com.example.ui.ui.screens.home.HomeViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
